rework of population raster script, in preparation to incorporate it into full pre processing as a step
