+++ config/database_config.ini
@@ -1,0 +1,15 @@
+[Input]
+databaseString:None
+database:arcgis
+user:postgres
+password:postgres
+host:None
+port:None
+resolution:100
+EPSGspatialReference:3035
+
+[Sql]
+x:x_mp_100m
+y:y_mp_100m
+value:"Einwohner"
+table:public."population"
