+++ src/util.py
@@ -102,76 +102,6 @@ def solve_iteratively(Config, region_octtree, regions, raster, raster_affine):
 
     return region_octtree
 
-def load_data2(Config, min_x, min_y, max_x, max_y):
-    database_string = Config.get("Input", "databaseString")
-    if database_string:
-        conn = psycopg2.connect(None, "arcgis", "postgres", "postgres")
-    else:
-        db = Config.get("Input", "database")
-        user = Config.get("Input", "user")
-        pw = Config.get("Input", "password")
-        host = Config.get("Input", "host")
-
-        conn = psycopg2.connect(database=db, user=user, password=pw, host=host)
-    cursor = conn.cursor()
-
-    sql = """SELECT {sql_x}, {sql_y}, {sql_value}
-            FROM {sql_table}
-            WHERE {sql_x} between %s and %s AND {sql_y} between %s and %s
-            """.format(sql_x = Config.get("Sql", "x"),
-                       sql_y = Config.get("Sql", "y"),
-                       sql_value = Config.get("Sql", "value"),
-                       sql_table = Config.get("Sql", "table"))
-
-    resolution = Config.getint("Input", "resolution")
-
-    print "parameters", (min_x, max_x, min_y, max_y)
-    cursor.execute(sql, (min_x, max_x, min_y, max_y)) #xmin xmax, ymin, ymax in that order
-    #ttes charhra
-
-    records = cursor.fetchall()
-    x_vals = zip(*records)[0]
-    y_vals = zip(*records)[1]
-    count_rows = len(set(y_vals))
-    count_cols = len(set(x_vals))
-    db_min_x = min(x_vals)
-    db_min_y = min(y_vals)
-
-    print (count_rows, count_cols, db_min_x, db_min_y)
-
-
-    pop_array = np.zeros((count_rows, count_cols), dtype=np.int32)
-
-
-    a = Affine(
-            resolution,
-            0,
-            db_min_x, #shift from the center marking to bottom left corner
-            0,
-            -resolution,
-            db_min_y + count_rows*resolution #shift from the center marking to bottom left corner
-    )
-
-    print "array origins: ", (db_min_x, db_min_y)
-
-
-    for line in records:
-        if line[2] > 0:
-            (x,y) = (line[0], line[1])
-            #print "(x,y): ", (x,y)
-            (col, row) = ~a * (x,y)
-            try:
-                pop_array[row, col] = line[2]
-            except IndexError:
-                print "(col,row): ", (col, row), "x,y:", (x,y)
-
-        #reference arrays by (row_no , col_no)
-        #reference arrays by (   a_y,      a_x   )
-
-    print np.sum(pop_array)
-
-    return (pop_array, a)
-
 def run_tabulate_intersection(zone_octtree, octtree_crs, land_use_folder, land_use_crs, class_field, field_values):
     print field_values
